deploy
I have also set environment variable JEKYLL_ENV=production in .bash_profile as if this is not done then it affects both the google seach code and the generation of sitemap.xml.

At some point I need to change that code and ensure it generates locally what I expect on the final machine.  This is because I build locally and then deploy to github.