I missed off the .html last time for comments if
I still don't know if this will actually work or not until I try it in
production
If it doesn't then I will probably just leave it as it is as I shouldn't
need to change it for future comments

Here is the problem.

I use to use Just Comments before they announced they were stopping.
They then sent a note advising Hyvor would accept their comments,and I
had 2 I wanted to keep.  However the default Hyvor Id was different to
what Just Comments used and so the new Hyvor Id couldn't find my two old
comments.

I prefer the Hyvor Id and want to use that on everything except the one
post with comments.

I believe if I do that I should see my old comments.  This may be a
falicy, and if it is I will just say goodbye to my comments