added status and status_date to post front matter
It now displays the value of the "status" front matter and then the
"status_date" front matter if it exits.

It takes a long time to write some posts.  I've decided to write so much
and then I can publish it with a status: of "WIP" for Work In Progress
or something else.  I can optionally add a status_date: to say when I
last updated.

I think I might use:
"Unfinished" when publishing a post that is not complete
"Updated" and a status_date: when I update a post, and leave the status:
blank when I think I've finished a post