standard libraries now gems and updated bundler
bundle exec jekyll serve was telling me two standard gems were not going
to be included with ruby releases in the future so they are now declared
gems in Gemspec.

Bundler was updated because I want to use latest versions if I can.  I
also used gem cleanup to remove the old versions.  Not to do with this
project but I wanted to ensure it isn't affected by that change.